thread: Fix broken assertion in Thread::migrate_away()
The assertion was incorrect, because the relevant `_pending_rqq` is that
of the thread's home CPU, which for a thread on an offline CPU isn't
the same as `_pending_rqq.current()`. Although being incorrect, the
assertion did not fail, because `Spin_lock::test()` is also fulfilled
if only the CPU lock is held. The proper test would have been
`Spin_lock::is_locked()`.

While at it, move the entire usage of `_pending_rqq.current()` into the
`!remote` if block to avoid confusion.

Change-Id: Iae5bd34819c2f5cb87dbd366aa8990717bb5c793