Properly block interrupts on the Local APIC level
Block interrupts on the Local APIC level by writing the value of 0xFF to
the Task Priority Register. This configures the interrupt priority
cutoff gate to class 15, subclass 15, blocking all interrupts with lower
or equal priority value (i.e. all of them).

The previous code was misguided by (probably) assuming that the cutoff
value affects the interrupt vectors or pins to block.

Change-Id: I2677b46214496b3c400e6f49e31355cf43dfaa19