Unify naming conventions related to interrupts
The verb "allocate" is generally used in connection with acquiring
memory resources. The methods that attach an Irq_base object to a
global system interrupt or an interrupt pin do not explicitly acquire
any memory resources in the common sense (although they might acquire
an interrupt vector on some platforms).

Therefore we replace the term "allocate" with the term "attach" in
these contexts to capture the semantics of the operation better.

For the sake of symmetry, we also rename the reverse operation from
"unbind" to "detach". This helps the disambiguation of the verb "bind"
(past tense "bound") which is used in several other contexts (e.g.
"IRQ being bound to a thread").

No change in functionality.

Change-Id: I7776259fb27c37eb85ee56b0285f19a2528021c1