Backoff after an unsucceful setting of target CPU of an IRQ
Despite not being implemented yet, setting a target CPU of an IRQ
involves a relatively costly search for an available interrupt vector
on IA-32 and AMD64 (especially in case the search is eventually
futile).

To avoid this repeated costly overhead, refrain from taking any action
for a certain number of attempts after a failed attempt to set the
target CPU of an IRQ.

This change only affects IA-32 and AMD64 at this moment since there is
no need for this mechanism on other platforms currently.

Change-Id: I8a13096754bf0a1b360dfe7252caa05f0bd1d25e