irq_sender: Fix sender list enqueue with disabled IRQ shortcut
Whether the IRQ sender needs to be enqueued into the sender list of the
receiver does not depend on whether the IRQ shortcut is taken (or even
enabled), but solely on whether the receiver is ready to immediately
receive the IRQ, as indicated by `receiver->sender_ok(this)`.

The refactor in dcfeb397842b4a85193db40d2dadd9e4a3d753e1 seems to be the
origin of the broken sender list enqueue with disabled IRQ shortcut. It
removed the unconditional sender enqueue in `Thread::check_sender()` but
not for the IRQ sender.

Fortunately, the Irq_shortcut option is always enabled on all relevant
architectures, only for sparc and ppc32 it is disabled, which explains
why it was never noticed.

Change-Id: I20872892c09abc4325c77a4f9ab0ef7ff7ab5c00