thread-ipc: Fix use-after-free in receive phase of closed wait
Under certain circumstances, a thread that performs a closed wait uses
a sender in its receive phase that has already been freed. The problem
is an insufficient check in Thread::get_next_sender(), that tries to
determine if a sender is enqueued in the sender list of the thread
doing the closed wait. Unfortunately that check is vulnerable to a
use-after-free if the sender already has been deleted in the meantime.

The conservative approach for the check would be to just iterate the
sender list and see if it contains the sender. However, that is
considered undesirable from a performance and a latency perspective
(think of a very long sender list).

See ticket #CD-1466.

Change-Id: I84a61179ad9dbd6addf31ee611a191378da0008d