ipc_gate: Abort blocked senders when IPC gate is destroyed
When an IPC gate is destroyed it unblocks all senders via unblock_all().
For the senders, it appears as if the IPC gate is now ready, they try to
access it, while in the meantime the IPC gate has already been freed.

The same can happen with a sender that is unblocked while the IPC gate
is still functional (not in destruction). Then the sender does not
execute for a long time, e.g. another higher prio thread hogs the CPU.
The IPC gate gets deleted. Then finally the sender executes again.

The core of the issue is that an unblocked sender has no way of knowing
whether the IPC gate it blocked on is still alive.

To allow the sender to determine whether the IPC gate is still alive,
when unblocking a sender the IPC gate moves the sender from the waiting
queue to the newly introduced unblocked queue. When the IPC gate is
destroyed, it clears both the waiting queue and the unblocked queue.

When the sender continues execution it has to check whether it is still
enqueued in the IPC gate's wait or unblocked queue. If not, it knows the
IPC gate has been deleted. Otherwise it dequeues itself from the queue
and does the send via the IPC gate.

See ticket #CD-1472.

Change-Id: I852be5b35f3360536cc5f8a601e18a9179d278ec