thread: Remove broken dequeue from sender list when killing thread
The sender list of a Receiver must only be accessed on its home CPU, the
lock of the Locked_prio_list is unused for Receiver::_sender_list.
Therefore, the killed thread cannot safely access it here.

At this point the killed thread must not be enqueued in any sender list,
because do_kill executing as as a continuation means that to end up here
the thread has completely unwound its stack. And that should include
aborting any ongoing IPC send operation or block operation on an
IPC gate or Semaphore.

So we cannot dequeue the killed thread, but that does not matter because
it must not be enqueued in any sender list anyway.

Change-Id: Ib37a83d6f3e8a714972e54ef6cb980f7124843a8