Separate Locked_prio_list and Iterable_prio_list
Previously Iterable_prio_list and Locked_prio_list where implemented the
same, with the latter being a typedef of the former. But they are used
in very different ways:
  - Iterable_prio_list as the sender list in Receiver, which must only
    be accessed on the Receiver's home CPU, the lock is unused.
  - Locked_prio_list as the waiting list in Ipc_gate and Semaphore, the
    lock is used, but _cursor is unused.

This is both confusing and wasteful. Also from an optimization
perspective, e.g. `Receiver::sender_list() == Sender::wait_queue()`
there seems to be no benefit in using the same type of prio list
everywhere.

So, separate the implementations of Locked_prio_list and
Iterable_prio_list, and change the type of Sender::_wq to the common
base type Prio_list.

Change-Id: I7a30ed5a3e204b35e332609bc6924ec0cc7b7f9c