cxx: dlist: avoid element type when manipulating list head
The list elements of a cxx::Dlist have the list pointers as members
(D_list_item). The list head is stored as a plain D_list_item, though.
When manipulating the list head, it is dangerous to cast to the list
element type because that is bigger and contemporary compilers (e.g.
gcc 12) will issue a warning along these lines:

  In static member function
          ‘static cxx::D_list_cyclic<T, C>::Iterator
          cxx::D_list_cyclic<T, C>::insert_before(T*, const Iterator&)
          [with T = ...; C = cxx::D_list_item_policy]’,
      inlined from ... at ...:
  ../include/l4/cxx/dlist:125:16: warning: array subscript 0 is outside
  array bounds of
      ‘L4Re::Blk::Brq_list [1]’ {aka ‘cxx::D_list<...> [1]’}
      [-Warray-bounds]
    125 |     C::next(e) = *pos;
        |     ~~~~~~~~~~~^~~~~~

Restructure the code to just work on the D_list_item type.

Sync with userland, Change-Id: I0b7997c0f9170f1f68cbcc91884798898d4e95da

Change-Id: I05fc97a4a310b0c446d8af8b9ffadae13c9d8dfa
Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>