Optimize Irq_chip_gen::irq(), Irq_chip_i8259_gen::irq()
These methods do not have to take the _lock spinlock and explicitly
check the _reserved bitmap to remain correct. This is possible thanks
to the fact that a non-null pointer can be stored in the _pins array
only if the respective bit field in _reserved is not set (which is
checked with the spinlock taken) and the store is atomic.

This improves interrupt processing latency and avoids a contention
point.

Change-Id: I252f6f97c3d965a04f59527613bf1915acf78179