Irq_sender: Improve handling of pending and revoked sends
When (un)binding the target of an Irq_sender the handling of the Pending
flag (A) and directly injected IRQs that are either revoked (B) or were
already pending again, was quite confusing, and slightly incorrect.

In particular, under certain circumstances, it might have been possible
that an Irq_sender gets enqueued twice:
1. In finish_replace_irq_thread() we clear the Pending flag or revoke a
   directly injected IRQ in Ok_was_pending state (finish_send() was
   already invoked for it, which cleared Queued).
2. IRQ hits concurrently, resulting in the Irq_sender getting enqueued
   at target thread.
3. Since the Irq_sender was pending in step 1, finish_replace_irq_send()
   now re-injects it, resulting in the Irq_sender, that is its Drq_item
   or Ipc_sender, getting enqueued a second time. This is broken.

Prevent this for (A) by not only clearing the Pending, but instead
transitioning from Pending->Queued.
Prevent this for (B) by first trying to requeue the send, i.e. try to
set the Queued flag, and only if that succeeds send to the target.

The underlying invariant, which was not upheld previously, is that in
order to send to a target the Queued flag must be set. And for each time
the Queued flag gets set, exactly one send must be made.

Further simplify the code, by relying on reject_send() to take care of
doing the re-inject. Conceptually, revoking/aborting a queued Irq_sender
can be considered a reject, not triggered by the receiver but by the
sender itself.

Change-Id: I6f85807eb7869cae91cb797db0ff419c6fa33340