Irq_sender: Separate Queued and Pending states
The Queued and Pending states were not cleanly separated, i.e. they were
not exclusive. It was possible for both Queued and Pending to be set.

That isn't a problem per se, given that when evaluating the Pending flag
in finish_replace_irq_thread(), we check whether queued is already set,
i.e. an IRQ that hit again while already marked pending, for example
because user-space unmasked the IRQ again.

However, this still is confusing, making it harder to reason over the
valid state transitions of Irq_sender. And since the additional checks
required to cleanly separate Queued and Pending do not add any
significant overhead, doing so seems like the right thing to do.

Note: Adding `!is_pending()` to `can_send()` in addition prevents
user-space from unmasking an IRQ that is in Pending state.

Change-Id: I1ff93428093d8cc5fb0c05f607b6d3e4b72e90d8