VMX: Remove Vm_vmx and Vm_vmx_t
Given the fact that Vm_vmx is inaccessible, remove that class and its
own methods.

Given the fact that the Vm_vmx_ept class is the only variant based on
the Vm_vmx_t template, remove the template plumbing entirely and move
the methods of the template to the Vm_vmx_base base class.

Except for temporarily making the methods load_vm_memory() and
store_vm_memory() virtual to allow the transition, this does not change
any functionality.

Change-Id: Ic9c884cf236d5c92eec855750dd4d01fa3329f98