arm: mention when a CPU boots but hide IDs information if already shown
Since Ib0b5dc687d9b86feb51ff2d1e9dc12974453dc75 the information that an
application CPU boots was hidden if its IDs are equal to the IDs of the
boot CPU to reduce logging for platforms with many CPUs.

This has the confusing side effect that for big.LITTLE platforms the CPU
IDs for the first group were shown for only one CPU (the boot CPU) and
the CPU IDs for the second group were shown for every CPU of that group.

Now we show at least one line for each booted CPU and determine if the
corresponding IDs were already shown.

Change-Id: Idc4e18cf23f5535e8373ce26250840b28a5b6d16