Modules.generic/coverage: fix clang version check
The version check only makes sense for llvm-based coverage, since only
there we need a compiler-specific library. For gcc it does not, so
only do the check and print the warning on llvm / clang builds.

Also, fix that we fall back to the maximum version present when we do
not know the compiler.

Change-Id: I7f66c4221db53eae9037565f5a479bb28ed937d5