utest: introduce Slab_store to avoid using static Kmem_slab_t instances
With CONFIG_AMP enabled we have to avoid using C++ singletons because
this feature always requires to have at least the guard variable to live
in .bss which is forbidden with CONFIG_AMP.

Introduce a simple storage for Kmem_slab_t instances stored in an array
as global data. So far the store is not yet used -- that will happen in
changes on top.

Change-Id: Ia8abfb789a6b95511406f55c5e9830fe87982dcf