Add capdb.yaml
This file describes the factory interface, the kernel factory provides.
It is used by L4ReStudio.

In other packages, the capdb also contains information on command line
parameters and more elaborate documentation. This is not needed for the
kernel, so the file is rather sparse. It also is not intended to
generate a `usage.md` file.

Change-Id: I2eaf5803fb1796b6ed38488c23d2726b511d5488