arm64: make no assumption about MPU on kernel re-entry
If the exception vector is entered from the same exception level, we
did assume that the MPU has the correct kernel state. While this is
usually correct, it is not universally true. For example, an "illegal
state" exception will keep the exception level but the MPU is already
in the "user space" state.

Hence, we must not assume anything about the MPU state. Always restore
the heap region to be on the safe side.

Change-Id: I0c34f2a8259ae01f11ad3bc7b6ef3b270c18459d