gcc_lib: fix __ashldi3() / __ashrdi3() for shifts >= 64
According to the C99 standard, the result of a shift operation is
undefined if the number of shifts is greater than or equal to the width
of the to-be-shifted value.

Hence, we must avoid shift operations with >= 32 bits on a 32-bit
operand.

Most likely this function is never called with bits >= 64 because the
compiler generates these calls only if there is no corresponding
hardware support. But the function definition does not restrict the
input range, so better play safe.

Change-Id: Ic3c57c22db951e3209dbc05956986b7f46e2b177