arm: cortex-r52: add L1 cache way segregation configuration
Introduce support for configuring L1 cache way segregation on Cortex-R52
for both instruction and data caches. This feature allows splitting cache
ways between the Flash and AXIM interfaces to optimize performance based
on memory access patterns.

Previously, this configuration was applied in both the kernel and
bootstrap platform init hooks. It is now centralized in the kernel to
maintain a single point of truth and simplify platform initialization.

Change-Id: I1014d6c2e67e47ac1fccfc6c295f8cc743785f22
Signed-off-by: Manuel Argüelles <manuel.arguelles@nxp.com>