make support for alien threads configurable by CONFIG_ALIEN
Makes it possible to disable the alien mode feature: The kernel produces
an exception IPC for each IPC and exception caused by a thread in alien
mode instead of handling these events regularly.

This feature can be used to attach a debugger to a thread and trace all
object invocations and their results. It could also be used to handle
other systems that use the same syscall instruction as L4Re.

This feature is only required for certain environments so it might make
sense to disable it normally.

This change is intended to improves code coverage.

Change-Id: I93f08ab96e41971808588b4153be2ffc03c4f733