arm32: initialize more system registers when enabling ext vCPU state
Also load extended vCPU state when enabling extended vCPU mode to load
certain system registers immediately, in particular the VBAR register.

On ARM32, the VBAR register is only part of the extended vCPU state. In
contrast, on ARM64, the VBAR register is part of the Context_hyp state
and will always be context-switched, not only when switching from/to a
context with extended vCPU mode enabled.

The reason is the HSTR register which on ARM32 prevents access to most
of the PL1 system registers for a context running in PL1 kernel mode.
This feature isn't available on ARM64, hence the different handling.

See ticket #CD-1573.

Change-Id: I9f04b83f5d32a4fd268ac237a561a9821232a18c