arm/64: remove Context_Hyp.{vmpidr,vpidr} and unify VMPIDR_EL2 init
The fields for both registers in Hyp_vm_state was never written but only
used as initial value for Hyp_vm_state.{vmpidr,vpidr}.

Unify the VMPIDR initialization in Thread::arch_init_vcpu_state(): Since
I7bc34a4aa2667154cd206455120d44495c556eb9 we used the MPIDR as initial
value for VMPIDR on ARM32. As the affinity values, the 'MT' bit and the
'U' bit are set by the VMM anyway, initialize these bits with zero. The
Multiprocessing Extension is mandatory if the processor implements the
Virtualization Extension, therefore always set the 'M' bit. As of ARMv8,
'M' is RES1.

Change-Id: Ifb5db3022eb0e91918bcfb02da477a4dda2c69bd