libc/stdio/vfprintf: don't use the same variable for different purposes
Introduce another variable 's_base' replacing 'a' for the first part of
the function. In the 'switch(t)' statement in the code below, 'a' is
initialized for every 'case' statement, so the value of 'a' from the
first part of the function (where it was replaced by 's_base') is not
used in the remaining function.

Make 's_base' a 'char const *' pointer because it's not used for
writing.

Change-Id: I15ebae9df2105c2c8795f233dee3c0bfd4beee4b