Makeconf: move -Map=fiasco.map from Makeconf to gen_kernel_rules
So far KERNEL_LDFLAGS contained "-Map=fiasco.map". This was problematic
because when linking kernel unit tests, a map file for a binary replaced
the map file for the binary which was generated right before so every
map file generate for a kunit test usually replaced fiasco.map.

Put the -Map=... command line parameter into the rule for generating the
kernel binary for being able to generate the map file exclusively for
the kernel. Generating a map file for every kunit test appears to be
wasted disk space.

Change-Id: I7f39603994ac44f5d63452215696701439953072