libc/stdio/vfprintf: simplify locking
The original musl libc implementation used FLOCK/FUNLOCK to protect the
_IO_FILE structure from concurrent access by different threads. Since
_IO_FILE is now always a local variable (only defined on the stack of
vsnprintf(), vprintf(), and printf()), this is no longer necessary.

We still need to prevent concurrent writing from different threads to
stdout. However, no locking is required for vsnprintf().

To simplify things, use the _IO_FILE::buf pointer to distinguish if
locking is required or not.

Change-Id: I0f47bea9768dcc730eb72a59520dfcf8284aee1b