x86: Remove INSTALLFILE_RULE mechanism
The INSTALLFILE_RULE mechanism was only present for ia32 and amd64.
It even is slightly broken, because the INSTALLFILE_RULE for the fiasco
binary is also invoked for each unit test.

We assume no one uses this feature, and thus remove it. If someone
actually used this mechanism, they can instead just run the install
commands externally after building fiasco.

Change-Id: Icab4e44d067459d49fac45c978d2d7bb928eb2ad