AVL set: Implement assignment operator, align copy constructor
Since the AVL set allocates and manages nodes internally, implementing
an assignment operator that works in the same way as the copy
constructor (i.e. deep-copying the nodes) is non-problematic. To that
end, we refactor the deep copy operation into a separate method.

Note that this change also changes the semantics of the copy
constructor to be aligned with the assignment operator: The newly
constructed object uses the allocator that it is specified with instead
of adopting the allocator of the source object. Arguably, this makes
more sense since the user is free to pick the desired allocator instead
of the source object overriding that choice.

Change-Id: I748ac1cbd3784907a93676ee29a5b0f1cd0868ed