jdb: more space for printing time stamps in trace buffer dump
When dumping the trace buffer (key 'D' while displaying trace buffer),
Jdb::write_tsc_s() prints the time stamps in the form 'S.UUUUUU s' where
'S' is the value in seconds and 'UUUUUU' is the fraction in microseconds
displayed as 6 digits. Before this change, the printed value was cut for
seconds > 9999 (4 digits). So for huge time stamps, not all 6 digits of
the microseconds part were shown.

Extend the space for displaying to allow printing seconds with up to 7
digits (sufficient for > 115 days) for the time stamp difference between
two events on the same CPU and up to 9 digits (sufficient for > 31.7
years) for displaying the absolute time stamp.

Also remove the precision format specifiers because the used strings
already have limited space.

Change-Id: Iee7f699ae168f0e948864a0ef8d9242c444e7df8