kobject_dbg: rename Get_name => Query_name and introduce new Get_name
Query_name is the better name for the existing Get_name function. It
actually takes a user-provided debug ID and returns the name for the
corresponding kernel object. This function now requires 'jdb' enabled
as well, 'rt_dbg' is not sufficient anymore.

Introduce Get_name which returns the name for a certain kernel object.
This function is available with 'rt_dbg'.

Change-Id: Id9d2f8490d805db36b25b1d694a61c8c1101dbad