ia32: remove Kmem::user_max()
Kmem::user_max() was only used at two places and was always set to ~0UL.
Originally this function was used to distinguish the maximum address for
small address spaces. This feature was removed long time ago.

Remove this function to reduce confusion as there are other similar
functions for other platforms.

Change-Id: Ib7694801d33ed7f3135e51b91ad89271a6523fd6