Irq_sender: Remove incorrect assertion
We must not assume that a kernel object that gets destructed has gone
through its destroy() method. The reason is that the initial map of the
object in `map_obj_initially()` can fail. In that case its constructor
is called immediately, which is fine since no one observed the object so
far, but still renders the assumption in ~Irq_sender() incorrect.

See ticket #CD-1440.

Change-Id: I12586b5dc4b4a85e9e16c55038d0a7c7f37dd7ca
Signed-off-by: Georg Kotheimer <georg.kotheimer@kernkonzept.com>