mem_space: obey max_user_address
The maximum available user space address depends sometimes on the
capabilities of the hardware. Use the actually supported value instead
of the theoretical page table maximum. This prevents user space from
mapping regions in the page table that are unusable by the underlying
hardware.

Except for arm64 cpu_virt, this has neither a functional nor a
performance difference.

Change-Id: Ib4896276ffe9ff81eb919e090e0046b9e2918fbc