Enforce minimal page size
Platforms register a set of possible page sizes. The current
implementation only enforces a descending order and a maximal
size of the page size added first - Page_order(Config::PAGE_SHIFT).
Adding a smaller page is silently accepted and leads to crashes
afterwards.

Here we add an assertion for the argument and modify the assertion
for the last page size to enforce a minimal page size and document
the underlying assumptions

Change-Id: Iebcec9bd0a8d3a177cd2783f10fdc0af88d39bac
Signed-off-by: Jean Wolter <jean.wolter@kernkonzept.com>