Improve readability of init_mapdb_mem()
The usage of the loop variable is pretty confusing and kind of
obfuscates the intention of the initialization and decrement at the end
of each loop.

Here we take a more direct approach:
  * start with the physical address bits,
  * decrement by Page_order(Max_mdb_ps_diff) on each loop and
  * pick the next entry of the global page size entries if it is larger
    than the current candidate.

The loop terminates when we reach the smallest possible page order
Page_order(Config::PAGE_SHIFT) which is always present.

The default value used in the original implementation is removed since
the loop was and is executed at least ones, since phys_bits is always
greater/equal Config::PAGE_SHIFT.

Change-Id: I5821a64ef2e0b699247e926716a5e20bdda65a4e
Signed-off-by: Jean Wolter <jean.wolter@kernkonzept.com>