Split Cpu::set_online() and Cpu::set_present()
Instead of calling set_online() and set_present() with a boolean
argument that could be quite unintuitive, split the methods into their
positive and negative variants that do not require any arguments and
should be self-explanatory:

* set_online(bool)  -> set_online(), set_offline()
* set_present(bool) -> set_present(), set_absent()

This should also ease searching for the call sites of the methods.
This does not cause any change of the observable behavior.

Signed-off-by: Martin Decky <martin.decky@kernkonzept.com>
Change-Id: I49b0a1e409415d6bda73ade46c368843dded2556