arm: clear VDISR_EL2 when switching to an extended vCPU context
The RAS extension is not supported by the kernel at the moment. But if
FEAT_RAS is implemented, a read or write access to DISR_EL1 from EL1
will cause a read/write access to VDISR_EL2 because HCR.AMO=1.

Therefore, on AArch32, VDISR_EL2 must be cleared when switching to an
extended vCPU context. Threads not running in extended vCPU context but
running at PL1 cannot access this register due to HSTR.T12=1.

On the other hand, on AArch64, VDISR_EL2 must be cleared when switching
to any thread running at EL1.

VDISR_EL2 is only present if FEAT_RAS is implemented. For now we assume
that even on big.LITTLE processors all cores implement this feature or
all cores do not implement this feature. Otherwise we panic.

See ticket #CD-1591.

Change-Id: Id7dbd963be4e93b17c984263d5c98b06123fc1d1
Signed-off-by: Frank Mehnert <frank.mehnert@kernkonzept.com>