Improve code robustness by removing pointless CPU arguments
Passing an argument that can only ever be set to a single logical value
(i.e. the current CPU) to a method is not just pointless, it also makes
the contract of the method vague and renders the call site needlessly
error-prone.

In this change, we make sure that methods that are only ever supposed
to work correctly with the current CPU are really just called from the
context of the current CPU. The burden of proof is typically naturally
carried by the caller.

No change in functionality.

Signed-off-by: Martin Decky <martin.decky@kernkonzept.com>
Change-Id: Ic12ffe9b77465c2e274918f13cc0d1e0feb40291