serial: introduce -serial_input command line parameter
In contrast to -serial_esc, this parameter enables input through Vlog
without interpreting escape sequences and without entering JDB when the
ESC key was pressed.

This is useful for setups where the escape key shall not enter JDB. As
another benefit, the kernel does no longer poll the UART for 10ms after
receiving ESC: This period is used to wait for any further keys in an
escape sequence.

Once JDB is entered, the filter console is enabled and JDB can handle
cursor key sequences.

Change-Id: I979749bbcd24b1dbf98d066cea9acb7def6dba05
Signed-off-by: Frank Mehnert <frank.mehnert@kernkonzept.com>