x86: Add support for tickless idle
On architectures, such as ARM or RISC-V, that do not need to open
interrupts to wait for the next interrupt, we can leave tickless idle
before executing the interrupt handler. That is impossible on x86,
since we need to open interrupts to wait for the next interrupt.
Then once an interrupt hits, execution continues in an interrupt
handler. Because the handler may schedule away from the idle thread
or perform operations where the CPU must not be in tickless idle,
the handler must leave tickless idle before handling the interrupt.

This unfortunately adds a small overhead to each interrupt entry,
because the interrupt handler always needs to check whether the CPU
is in tickless idle.

Co-authored-by: Christian Ludwig <cludwig@genua.de>
Co-authored-by: Frank Mehnert <frank.mehnert@kernkonzept.com>
Change-Id: I21309b13e473be642624e4ad79bdecfee39a6709
Signed-off-by: Georg Kotheimer <georg.kotheimer@kernkonzept.com>