arm/64: play safe and use 'asm volatile()' instead of 'asm()'
Use 'asm volatile()' for code which is executed as part of an if/else
statement and which accesses system registers and where this access must
only be performed if the corresponding condition says so.

In this specific example the compiler is free to read all SCTLR system
registers (SCTLR_EL1, SCTLR_EL2, SCTLR_EL3) and just discard the unused
values.

Although we did not spot access to the wrong SCTLR register in generated
code, play safe here by using 'asm volatile'.

Change-Id: I9b1cac3be4d53140167ff8477a82143f4b755fe0
Signed-off-by: Frank Mehnert <frank.mehnert@kernkonzept.com>