arm64: Allow 3-level and 4-level page-table in parallel
Make 3-level and 4-level page table available in parallel, thus allowing
to detect which page-table to use at runtime and thus no need to
require to build the kernel for a specific processor variant.
For specific cases we provide a configuration option to select one
implemention out of the currently two available, reducing text segment
requirements.

Change-Id: Icca3caaa10c3c55a4455fe8cd016854d815ddea0
Signed-off-by: Adam Lackorzynski <adam@l4re.org>