mmu-arm-32: fixed usage of 'asm volatile' instead of 'asm'
In certain cases it is important to use 'asm volatile (...)' and not
only 'asm (...). With the latter variant, the compiler is free to
execute the assembler statement even if the result is not used. The
former variant is more strict.

In this specific case the CCSIDR2 system register must only be read if
FEAT_CCIDX is implemented, otherwise the access is undefined. In our
case, the resulting exception was not handled and the kernel just hang
very early during boot.

With 'asm' the compiler generates code for accessing CCSIDR2 in certain
cases _unconditionally_ and just discards the resulting value. With 'asm
volatile', the CCSIDR2 read access is only performed if the if-statement
returns true -- which is the required behavior.

See ticket #CD-1648.

Change-Id: I27a2f4df3cc258e817c24a52f097d036a8e42149
Signed-off-by: Frank Mehnert <frank.mehnert@kernkonzept.com>