MIPS: fix atomic_add_fetch() inline assembler earlyclobber operand
The result needs to be marked as 'earlyclobber' because it is written
before the code has finished using input operands.

In certain cases, the register which was assigned to 'ptr' was also
assigned to 'res' so the 'move' instruction was actually overwriting
the target address of the 'sc' instruction and the atomic instruction
did never succeed.

See ticket #CD-1557.

Change-Id: I9b1d9555059ee9e2b9b135902fd004b00eff3bf6
Signed-off-by: Frank Mehnert <frank.mehnert@kernkonzept.com>