VMX: Make the calls of enforce() and allowed() explicit
Instead of relying on a default value of the argument, always specify
explicitly whether we enforce() the true or the false value of the
given bit. Split the original allowed() method into allowed() and
disallowed().

These modifications should make the call sites of the methods more
transparent and less error-prone.

While at it, refactor the body of Bit_defs_t::relax() into
Bit_defs::relax_bits() for consistency.

This commit does not change any functionality.

Signed-off-by: Martin Decky <martin.decky@kernkonzept.com>
Change-Id: If63bbde69467bd95149f80dacb07690424671795